Got no_buffered reads working

+++ GraphReLabel/InvertAndRelabelNodes.h
@@ -7,8 +7,8 @@ template <typename T1, typename T2>
 class InvertAndRelabelNodes
 {
 public:
-    uint32 total_read;
-    uint32 total_write;
+    uint64 total_read;
+    uint64 total_write;
 
     GraphReader<T1, T2>* graph;
     //HashManager* hashManager;
